requirements.txt
asgiref==3.3.4
Django==3.2.2
djangorestframework==3.12.4
pytz==2021.1
selenium==3.141.0
sqlparse==0.4.1
urllib3==1.26.4
